Restructuring inputs directory
