+++ src/lava/lib/dnf/inputs/gauss_pattern/process.py
