+++ tests/lava/lib/dnf/unit/inputs/gauss_pattern/__init__.py
