+++ tests/lava/lib/dnf/unit/inputs/gauss_pattern/test_process.py
