--- tests/lava/lib/dnf/unit/inputs/patterns/gauss_pattern/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (C) 2021 Intel Corporation
-# SPDX-License-Identifier: BSD-3-Clause
-# See: https://spdx.org/licenses/
-
-__import__("pkg_resources").declare_namespace(__name__)
