+++ tests/lava/lib/dnf/unit/inputs/spike_generator/test_process.py
